feat: fix durable state not using schemaName set in config correctly
